wip: issue 298 persistent client profiles
